agregando estilo a boton logout y logo de waiterProfile
