+++ bq/src/styles/App.scss
@@ -3,7 +3,10 @@
 @import "../styles/variables"; 
 html,
 body {
-  background-image: url("../assets/background.jpg");
+  background-image: linear-gradient(
+    rgba(0, 0, 0, 0.6),
+    rgba(0, 0, 0, 0.6)
+  ), url("../assets/background.jpg");
   margin: 0;
   padding: 0;
   font-family: 'Roboto', sans-serif;
