+++ bq/src/styles/_main.scss
@@ -2,3 +2,4 @@
 @import "../components/Login/Login.scss";
 @import "../components/WaiterProfile/WaiterProfile.scss";
 @import "../styles/App.scss";
+@import "../components/NavBar.jsx/NavBar.scss";
