+++ bq/src/styles/_variables.scss
@@ -1,9 +1,10 @@
-$primary: #e5dcc3;
+$primary: #FAEEE0;
 $primary-light: #fffff6;
 $primary-dark: #b3aa93;
-$secondary: #3d0000;
-$secondary-light: #6c2b2c;
+$secondary: #FF4848;
+$secondary-light: #FCD8D4;
 $secondary-dark: #240000;
 $text-on-primary: #FEF9EF;
 $text-on-secondary: #3F0713;
 $line-input: #E5DCC3;
+
